requirements.txt
Flask
mysql-connector-python
dicttoxml
PyJWT
